moved Tabs class from main.py to gui/bind/tabs.py
