+++ gui/bind/__init__.py
@@ -1,1 +1,2 @@
 from . import directory_tree, footer
+from .tabs import Tabs
